+++ luthier/src/main/java/com/yahoo/bard/webservice/data/config/luthier/LuthierIndustrialPark.java
@@ -4,20 +4,20 @@ package com.yahoo.bard.webservice.data.config.luthier;
 
 import com.yahoo.bard.webservice.data.config.ConfigurationLoader;
 import com.yahoo.bard.webservice.data.config.ResourceDictionaries;
-import com.yahoo.bard.webservice.data.config.luthier.factories.AggregationAverageMakerFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.ArithmeticMakerFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.CountMakerFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.DefaultLogicalTableGroupFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.DefaultMetricFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.DoubleSumMakerFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.KeyValueStoreDimensionFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.LongSumMakerFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.LuceneSearchProviderFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.MapKeyValueStoreFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.NoOpSearchProviderFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.PermissivePhysicalTableFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.ScanSearchProviderFactory;
-import com.yahoo.bard.webservice.data.config.luthier.factories.StrictPhysicalTableFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.metricmaker.AggregationAverageMakerFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.metricmaker.ArithmeticMakerFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.metricmaker.CountMakerFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.logicaltable.DefaultLogicalTableGroupFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.logicalmetric.DefaultMetricFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.metricmaker.DoubleSumMakerFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.dimension.KeyValueStoreDimensionFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.metricmaker.LongSumMakerFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.searchprovider.LuceneSearchProviderFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.keyvaluestore.MapKeyValueStoreFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.searchprovider.NoOpSearchProviderFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.physicaltable.PermissivePhysicalTableFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.searchprovider.ScanSearchProviderFactory;
+import com.yahoo.bard.webservice.data.config.luthier.factories.physicaltable.StrictPhysicalTableFactory;
 import com.yahoo.bard.webservice.data.config.luthier.table.LogicalTableGroup;
 import com.yahoo.bard.webservice.data.config.metric.makers.AggregationAverageMaker;
 import com.yahoo.bard.webservice.data.config.metric.makers.ArithmeticMaker;
