+++ luthier/src/main/java/com/yahoo/bard/webservice/data/config/luthier/factories/logicaltable/DefaultLogicalTableGroupFactory.java
@@ -1,6 +1,6 @@
 // Copyright 2019 Oath Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.data.config.luthier.factories;
+package com.yahoo.bard.webservice.data.config.luthier.factories.logicaltable;
 
 import com.yahoo.bard.webservice.application.luthier.LuthierConfigNode;
 import com.yahoo.bard.webservice.data.config.luthier.Factory;
