+++ luthier/src/main/java/com/yahoo/bard/webservice/data/config/luthier/factories/physicaltable/PermissivePhysicalTableFactory.java
@@ -1,9 +1,11 @@
 // Copyright 2019 Oath Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.data.config.luthier.factories;
+package com.yahoo.bard.webservice.data.config.luthier.factories.physicaltable;
 
 import com.yahoo.bard.webservice.application.luthier.LuthierConfigNode;
 import com.yahoo.bard.webservice.data.config.luthier.LuthierIndustrialPark;
+import com.yahoo.bard.webservice.data.config.luthier.factories.physicaltable.LuthierPhysicalTableParams;
+import com.yahoo.bard.webservice.data.config.luthier.factories.physicaltable.SingleDataSourcePhysicalTableFactory;
 import com.yahoo.bard.webservice.table.ConfigPhysicalTable;
 import com.yahoo.bard.webservice.table.PermissivePhysicalTable;
