+++ luthier/src/test/groovy/com/yahoo/bard/webservice/data/config/luthier/LuthierIndustrialParkSpec.groovy
@@ -1,13 +1,10 @@
 // Copyright 2019 Oath Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.data.config.luthier;
+package com.yahoo.bard.webservice.data.config.luthier
 
-import com.yahoo.bard.webservice.data.config.luthier.ConceptType
-import com.yahoo.bard.webservice.data.config.luthier.Factory
-import com.yahoo.bard.webservice.data.config.luthier.LuthierIndustrialPark
+
 import com.yahoo.bard.webservice.data.config.luthier.dimension.LuthierDimensionField
-import com.yahoo.bard.webservice.data.config.luthier.LuthierResourceDictionaries
-import com.yahoo.bard.webservice.data.config.luthier.factories.KeyValueStoreDimensionFactory
+import com.yahoo.bard.webservice.data.config.luthier.factories.dimension.KeyValueStoreDimensionFactory
 import com.yahoo.bard.webservice.data.dimension.Dimension
 import com.yahoo.bard.webservice.data.dimension.KeyValueStore
 import com.yahoo.bard.webservice.data.dimension.SearchProvider
